lc0582: Kill Process
