+++ src/main/java/com/zea7ot/leetcode/logs.txt
@@ -27,7 +27,7 @@ new
 5: 
 
 review
-3: 0436
+3: 0436, 0582
 4: 
 5:
