+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/breath_first_search/array_1d.txt
@@ -9,6 +9,7 @@ References:
 2   0017    https://leetcode.com/problems/letter-combinations-of-a-phone-number/
 2   0784    https://leetcode.com/problems/letter-case-permutation/
 
+3   0582    https://leetcode.com/problems/kill-process/
 3   1376    https://leetcode.com/problems/time-needed-to-inform-all-employees/
 
 4   0839    https://leetcode.com/problems/similar-string-groups/
