+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/breath_first_search/graph/graph_undirected.txt
@@ -11,6 +11,7 @@ References:
 
 3   0133    https://leetcode.com/problems/clone-graph/
 3   0323    https://leetcode.com/problems/number-of-connected-components-in-an-undirected-graph/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0785    https://leetcode.com/problems/is-graph-bipartite/
 3   0841    https://leetcode.com/problems/keys-and-rooms/
 3   0886    https://leetcode.com/problems/possible-bipartition/
