+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/breath_first_search/graph/summary.txt
@@ -13,6 +13,7 @@ References:
 3   0133    https://leetcode.com/problems/clone-graph/
 3   0210    https://leetcode.com/problems/course-schedule-ii/
 3   0323    https://leetcode.com/problems/number-of-connected-components-in-an-undirected-graph/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0743    https://leetcode.com/problems/network-delay-time/
 3   0785    https://leetcode.com/problems/is-graph-bipartite/
 3   0841    https://leetcode.com/problems/keys-and-rooms/
