+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/breath_first_search/summary.txt
@@ -38,6 +38,7 @@ References:
 3   0286    https://leetcode.com/problems/walls-and-gates/
 3   0323    https://leetcode.com/problems/number-of-connected-components-in-an-undirected-graph/
 3   0444    https://leetcode.com/problems/sequence-reconstruction/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0623    https://leetcode.com/problems/add-one-row-to-tree/
 3   0662    https://leetcode.com/problems/maximum-width-of-binary-tree/
 3   0743    https://leetcode.com/problems/network-delay-time/
