+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/breath_first_search/tree.txt
@@ -26,6 +26,7 @@ References:
 
 3   0101    https://leetcode.com/problems/symmetric-tree/
 3   0103    https://leetcode.com/problems/binary-tree-zigzag-level-order-traversal/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0623    https://leetcode.com/problems/add-one-row-to-tree/
 3   0662    https://leetcode.com/problems/maximum-width-of-binary-tree/
 3   1245    https://leetcode.com/problems/tree-diameter/
