+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/depth_first_search/array_1d.txt
@@ -21,6 +21,7 @@ References:
 3   0377    https://leetcode.com/problems/combination-sum-iv/
 3   0491    https://leetcode.com/problems/increasing-subsequences/
 3   0494    https://leetcode.com/problems/target-sum/
+3   0582    https://leetcode.com/problems/kill-process/
 3   1048    https://leetcode.com/problems/longest-string-chain/
 3   1273    https://leetcode.com/problems/delete-tree-nodes/
 3   1376    https://leetcode.com/problems/time-needed-to-inform-all-employees/
