+++ src/main/java/com/zea7ot/leetcode/tags/algorithm/search/depth_first_search/tree/summary.txt
@@ -62,6 +62,7 @@ References:
 3   0536    https://leetcode.com/problems/construct-binary-tree-from-string/
 3   0543    https://leetcode.com/problems/diameter-of-binary-tree/
 3   0545    https://leetcode.com/problems/boundary-of-binary-tree/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0623    https://leetcode.com/problems/add-one-row-to-tree/
 3   0652    https://leetcode.com/problems/find-duplicate-subtrees/
 3   0654    https://leetcode.com/problems/maximum-binary-tree/
