+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/array_1d.txt
@@ -75,6 +75,7 @@ References:
 3   0525    https://leetcode.com/problems/contiguous-array/
 3   0548    https://leetcode.com/problems/split-array-with-equal-sum/
 3   0560    https://leetcode.com/problems/subarray-sum-equals-k/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0611    https://leetcode.com/problems/valid-triangle-number/
 3   0659    https://leetcode.com/problems/split-array-into-consecutive-subsequences/
 3   0714    https://leetcode.com/problems/best-time-to-buy-and-sell-stock-with-transaction-fee/
