+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/array/summary.txt
@@ -95,6 +95,7 @@ This file serves as the total list of problems related to Array, regardless of a
 3   0523    https://leetcode.com/problems/continuous-subarray-sum/
 3   0547    https://leetcode.com/problems/friend-circles/
 3   0548    https://leetcode.com/problems/split-array-with-equal-sum/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0611    https://leetcode.com/problems/valid-triangle-number/
 3   0638    https://leetcode.com/problems/shopping-offers/
 3   0659    https://leetcode.com/problems/split-array-into-consecutive-subsequences/
