+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/tree/depth_first_search.txt
@@ -40,6 +40,7 @@ References:
 3   0530    https://leetcode.com/problems/minimum-absolute-difference-in-bst/
 3   0543    https://leetcode.com/problems/diameter-of-binary-tree/
 3   0545    https://leetcode.com/problems/boundary-of-binary-tree/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0623    https://leetcode.com/problems/add-one-row-to-tree/
 3   0652    https://leetcode.com/problems/find-duplicate-subtrees/
 3   0662    https://leetcode.com/problems/maximum-width-of-binary-tree/
