+++ src/main/java/com/zea7ot/leetcode/tags/data_structure/tree/traversal/summary.txt
@@ -16,6 +16,7 @@ References:
 2   0501    https://leetcode.com/problems/find-mode-in-binary-search-tree/
 2   0563    https://leetcode.com/problems/binary-tree-tilt/
 2   0589    https://leetcode.com/problems/n-ary-tree-preorder-traversal/
+2   0637    https://leetcode.com/problems/average-of-levels-in-binary-tree/
 2   0655    https://leetcode.com/problems/print-binary-tree/
 2   0669    https://leetcode.com/problems/trim-a-binary-search-tree/
 2   0701    https://leetcode.com/problems/insert-into-a-binary-search-tree/
@@ -41,8 +42,8 @@ References:
 3   0536    https://leetcode.com/problems/construct-binary-tree-from-string/
 3   0543    https://leetcode.com/problems/diameter-of-binary-tree/
 3   0545    https://leetcode.com/problems/boundary-of-binary-tree/
+3   0582    https://leetcode.com/problems/kill-process/
 3   0623    https://leetcode.com/problems/add-one-row-to-tree/
-2   0637    https://leetcode.com/problems/average-of-levels-in-binary-tree/
 3   0652    https://leetcode.com/problems/find-duplicate-subtrees/
 3   0654    https://leetcode.com/problems/maximum-binary-tree/
 3   0662    https://leetcode.com/problems/maximum-width-of-binary-tree/
